+++ packages/components/src/test/data/constants.ts
@@ -15,8 +15,7 @@
  */
 import { List, Map } from 'immutable';
 
-import { AssayWizardModel } from '../../internal/components/assay/models';
-
+import { AssayWizardModel } from '../../internal/components/assay/AssayWizardModel';
 import { QueryInfo } from '../../internal/components/base/models/QueryInfo';
 import { AssayDefinitionModel, AssayDomainTypes } from '../../internal/components/base/models/model';
 import { IFile } from '../../internal/components/files/models';
